requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.27.0
torch==2.9.1
transformers==4.57.3
scikit-learn==1.8.0
joblib==1.5.3
numpy==2.3.5
pandas==2.3.3
requests==2.32.5
python-dotenv==1.2.1
python-multipart==0.0.6
pydantic==2.12.5
httpx==0.28.1
asyncio==3.4.3
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
redis==5.0.1
celery==5.3.4